new home of structural validation in tests
